perf(部门): 添加删除部门用户存在校验

+++ carrot/src/main/java/com/carrothole/carrot/controller/AuDeptController.java
@@ -3,7 +3,9 @@ package com.carrothole.carrot.controller;
 import com.carrothole.carrot.authorization.PreAuthorize;
 import com.carrothole.carrot.entity.vo.PageVO;
 import com.carrothole.carrot.exception.ParamException;
+import com.carrothole.carrot.exception.UnSupportOperationException;
 import com.carrothole.carrot.property.CarrotProperty;
+import com.carrothole.carrot.service.AuDeptUserService;
 import com.mybatisflex.core.paginate.Page;
 import com.mybatisflex.core.query.QueryWrapper;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -23,6 +25,8 @@ import io.swagger.v3.oas.annotations.Parameter;
 import java.util.List;
 
 import static com.carrothole.carrot.entity.table.AuDeptTableDef.AU_DEPT;
+import static com.carrothole.carrot.entity.table.AuDeptUserTableDef.AU_DEPT_USER;
+import static com.carrothole.carrot.entity.table.AuUserTableDef.AU_USER;
 
 /**
  *  控制层。
@@ -37,6 +41,9 @@ public class AuDeptController {
 
     @Autowired
     private AuDeptService auDeptService;
+
+    @Autowired
+    private AuDeptUserService auDeptUserService;
 
     @Autowired
     private CarrotProperty carrotProperty;
@@ -67,6 +74,10 @@ public class AuDeptController {
     @Operation(description="根据主键")
     @PreAuthorize(menu = {"au:dept:remove"}, user = "carrot")
     public boolean remove(@PathVariable @Parameter(description="主键")String id) {
+        // 校验部门下是否有用户
+        if (auDeptUserService.exists(QueryWrapper.create().and(AU_DEPT_USER.USER_ID.eq(id)))){
+            throw new UnSupportOperationException("删除失败：部门下存在用户");
+        }
         return auDeptService.removeById(id);
     }
 
@@ -82,7 +93,7 @@ public class AuDeptController {
     public boolean update(@RequestBody @Parameter(description="主键")AuDept auDept) {
         // 校验同级下是否有同名部门
         if (auDeptService.exists(QueryWrapper.create().and(AU_DEPT.DEPT_NAME.eq(auDept.getDeptName())).and(AU_DEPT.ID.ne(auDept.getId())))) {
-            throw new ParamException("同级下已存在同名部门");
+            throw new UnSupportOperationException("创建失败：同级下已存在同名部门");
         }
         return auDeptService.updateById(auDept);
     }
